+++ assets/js/menu/parent-menu-icon.ts
@@ -8,10 +8,9 @@ export class ParentMenuIcon extends Base {
     }
 
     private add(): void {
-        this._j('.menu li > ul').before(
-            `<button class="${this._submenu_button_class_name} sub-menu-toggle">
-                ${this.renderCaret('down')}
-            </button>`
+        this._j('.menu li > ul').prev('a').append(
+            `<span class="${this._submenu_button_class_name}
+                sub-menu-toggle">${this.renderCaret('down')}</span>`
         )
     }
 }